[chore #159374956] -refactor as per LFA Tina's feedback: - remove challenge 1 and 2 files - make more oop
