--- UI/about.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>MyDiary</title>
-    <link href='https://fonts.googleapis.com/css?family=Titillium+Web:400,300,600' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" href="css/normalize.min.css">
-    <link rel="stylesheet" href="css/style.css">
-    <script src="js/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="js/navbar.js" type="text/javascript"></script>
-</head>
-
-<body background="images/bg.png">
-
-<!--Navigation Bar-->
-<div id='cssmenu'>
-    <ul>
-        <li><a href='#'><span>MyDiary</span></a></li>
-        <li><a href='entries.html'><span>Entries</span></a></li>
-        <li><a href='profile.html'><span>Profile</span></a></li>
-        <li class='active'><a href='about.html'><span>about</span></a></li>
-    </ul>
-</div>
-
-<div class="about">
-    <h1 id="aheading">MyDiary</h1>
-    <h3>This is an Online Journal That Allows You To Pen Down Your Thoughts & Feelings Freely and For Free</h3>
-    <h4>Version 1.0</h4>
-    <h4>Author: Byron Taaka</h4>
-    <h4>4dbyron@gmail.com</h4>
-
-    <h2><a href="https://www.pivotaltracker.com/n/projects/2184264">PivotalTracker Link</a></h2>
-</div>
-<script src='js/jquery.min.js'></script>
-<script src="js/index.js"></script>
-
-</body>
-</html>
