--- UI/css/buttons.css
@@ -1,53 +1,0 @@
-.button, #btnProfPic, #btnPassword {
-    border-top: 1px solid #96d1f8;
-    background: #65b5e6;
-    background: -webkit-gradient(linear, left top, left bottom, from(#0927eb), to(#65b5e6));
-    background: -webkit-linear-gradient(top, #0927eb, #65b5e6);
-    background: -moz-linear-gradient(top, #0927eb, #65b5e6);
-    background: -ms-linear-gradient(top, #0927eb, #65b5e6);
-    background: -o-linear-gradient(top, #0927eb, #65b5e6);
-    padding: 6.5px 13px;
-    -webkit-border-radius: 26px;
-    -moz-border-radius: 26px;
-    border-radius: 26px;
-    -webkit-box-shadow: rgba(0, 0, 0, 1) 0 1px 0;
-    -moz-box-shadow: rgba(0, 0, 0, 1) 0 1px 0;
-    box-shadow: rgba(0, 0, 0, 1) 0 1px 0;
-    text-shadow: rgba(0, 0, 0, .4) 0 1px 0;
-    color: white;
-    font-size: 17px;
-    font-family: Georgia, serif;
-    text-decoration: none;
-    vertical-align: middle;
-}
-
-.button:hover {
-    border-top-color: #4e39db;
-    background: #4e39db;
-    color: #f7d70a;
-}
-
-.button:active {
-    border-top-color: #1b435e;
-    background: #1b435e;
-}
-
-.btnCenter{
-    margin: 40%;
-    margin-right: 0%;
-}
-.btnEnd{
-    margin: 60%;
-    margin-right: 0%;
-
-}
-.btnSpaced{
-    margin: 22%;
-    margin-right: 0%;
-
-}
-
-.notifySpace{
-    margin: 12%;
-    margin-right: 0%;
-}
