--- UI/css/entries.css
@@ -1,64 +1,0 @@
-body {
-    position: relative;
-}
-
-.DivTable {
-    display: table;
-    position: absolute;
-    left: 5%;
-    background-color: white;
-
-}
-
-.Title {
-    display: table-caption;
-    text-align: center;
-    font-weight: bold;
-    font-size: 2rem
-}
-
-.Heading {
-    display: table-row;
-    font-weight: bold;
-    text-align: center;
-    background-color: lightblue;
-}
-
-.Row {
-    display: table-row;
-}
-
-.Cell {
-    display: table-cell;
-    border: thin solid;
-    padding-left: 5px;
-    padding-right: 5px;
-}
-
-#EntryView {
-    left: 30%;
-    right: 30%;
-}
-
-#EntryModify {
-    left: 30%;
-    right: 30%;
-}
-
-#subtitle {
-    display: marker;
-    text-align: center;
-    font-weight: bold;
-    font-size: 2rem
-}
-
-#EntryCategory {
-    font-size: small;
-    text-align: end;
-    color: navy;
-}
-
-.DivDel {
-    position: relative;
-}
-
