--- UI/css/profile.css
@@ -1,119 +1,0 @@
-
-#DivContainer {
-    width: 40%;
-    height: 80%;
-    padding-top: 60px;
-    border: 0px;
-    margin-left: 30%;
-    position: relative;
-    bottom: 0%;
-    background-color: aliceblue;
-
-}
-
-#UserProf {
-    width: 49%;
-    height: 65%;
-    border: 0px;
-    float: left;
-    margin-left: 0.5%;
-    text-align: center;
-}
-
-#userDetails {
-    width: 49%;
-    height: 65%;
-    border: 0px;
-    float: right;
-    margin-left: 0%;
-    margin-right: 0.5%
-}
-
-#userSettings {
-    width: 100%;
-    height: 32%;
-    padding-left: 15%;
-    clear: both;
-    text-align: left;
-
-}
-
-.ProfileDiv {
-    position: relative;
-}
-
-.ProfileDiv button {
-    position: absolute;
-    bottom: 0;
-}
-
-#btnPassword {
-    left: 0%;
-}
-
-#btnProfPic {
-    left: 25%;
-}
-
-img {
-    width: 80%;
-    height: 65%;
-}
-
-.notify {
-    position: relative;
-    display: inline-block;
-    width: 60px;
-    height: 34px;
-}
-
-.notify input {
-    display: none;
-}
-
-.reminder {
-    position: absolute;
-    cursor: pointer;
-    top: 0;
-    left: 0;
-    right: 0;
-    bottom: 0;
-    background-color: #ccc;
-    -webkit-transition: .4s;
-    transition: .4s;
-}
-
-.reminder:before {
-    position: absolute;
-    content: "";
-    height: 26px;
-    width: 26px;
-    left: 4px;
-    bottom: 4px;
-    background-color: white;
-    -webkit-transition: .4s;
-    transition: .4s;
-}
-
-input:checked + .reminder {
-    background-color: #2196F3;
-}
-
-input:focus + .reminder {
-    box-shadow: 0 0 1px #2196F3;
-}
-
-input:checked + .reminder:before {
-    -webkit-transform: translateX(26px);
-    -ms-transform: translateX(26px);
-    transform: translateX(26px);
-}
-
-/* Rounded sliders */
-.reminder.round {
-    border-radius: 34px;
-}
-
-.reminder.round:before {
-    border-radius: 50%;
-}
