--- UI/css/style.css
@@ -1,266 +1,0 @@
-*, *:before, *:after {
-  box-sizing: border-box;
-}
-
-html {
-  overflow-y: scroll;
-}
-
-body {
-  background-image: url("../images/bg.png");
-  font-family: 'Titillium Web', sans-serif;
-}
-
-a {
-  text-decoration: none;
-  color: #1ab188;
-  transition: .5s ease;
-}
-a:hover {
-  color: #0f71ba;
-}
-
-.form {
-  background: rgba(19, 35, 47, 0.9);
-  padding: 2%;
-  max-width: 600px;
-  margin: 10px auto;
-  border-radius: 4px;
-  box-shadow: 0 4px 10px 4px rgba(19, 35, 47, 0.3);
-}
-
-.tab-group {
-  list-style: none;
-  padding: 0;
-  margin: 0 0 40px 0;
-}
-.tab-group:after {
-  content: "";
-  display: table;
-  clear: both;
-}
-.tab-group li a {
-  display: block;
-  text-decoration: none;
-  padding: 10px;
-  background: rgba(160, 179, 176, 0.25);
-  color: #a0b3b0;
-  font-size: 20px;
-  float: left;
-  width: 50%;
-  text-align: center;
-  cursor: pointer;
-  transition: .5s ease;
-}
-.tab-group li a:hover {
-  background: #0f71ba;
-  color: #ffffff;
-}
-.tab-group .active a {
-  background: #0f71ba;
-  color: #ffffff;
-}
-
-.tab-content > div:last-child {
-  display: none;
-}
-
-h1 {
-  text-align: center;
-  color: #ffffff;
-  font-weight: 500;
-  margin: 0 0 15px;
-}
-
-label {
-  position: absolute;
-  -webkit-transform: translateY(6px);
-          transform: translateY(6px);
-  left: 13px;
-  color: rgba(255, 255, 255, 0.5);
-  transition: all 0.25s ease;
-  -webkit-backface-visibility: hidden;
-  pointer-events: none;
-  font-size: 22px;
-}
-label .req {
-  margin: 2px;
-  color: #1ab188;
-}
-
-label.active, .activated {
-  -webkit-transform: translateY(-20px);
-          transform: translateY(-20px);
-  left: 2px;
-  font-size: 14px;
-}
-label.active .req {
-  opacity: 0;
-}
-
-label.highlight {
-  color: #ffffff;
-}
-
-input, textarea {
-  font-size: 22px;
-  display: block;
-  width: 100%;
-  height: 100%;
-  padding: 5px 10px;
-    background: none;
-    border: 1px solid #a0b3b0;
-  color: #ffffff;
-  border-radius: 0;
-  transition: border-color .25s ease, box-shadow .25s ease;
-}
-input:focus, textarea:focus {
-  outline: 0;
-  border-color: #1ab188;
-}
-
-textarea {
-  border: 2px solid #a0b3b0;
-  resize: vertical;
-}
-
-.field-wrap {
-  position: relative;
-  margin-bottom: 30px;
-}
-
-.top-row:after {
-  content: "";
-  display: table;
-  clear: both;
-}
-.top-row > div {
-  float: left;
-  width: 48%;
-  margin-right: 4%;
-}
-.top-row > div:last-child {
-  margin: 0;
-}
-
-.forgot {
-  margin-top: -20px;
-  text-align: right;
-}
-
-@import url(http://fonts.googleapis.com/css?family=Capriola);
-#cssmenu,
-#cssmenu ul,
-#cssmenu ul li,
-#cssmenu ul li a {
-  margin: 0;
-  padding: 0;
-  border: 0;
-  list-style: none;
-  line-height: 1;
-  display: block;
-  position: relative;
-  font-family: Capriola, Helvetica, sans-serif;
-}
-#cssmenu {
-  width: auto;
-  height: 50px;
-  padding-bottom: 4px;
-  clear: both;
-}
-#cssmenu.align-right {
-  float: right;
-}
-#cssmenu.align-right ul li {
-  float: right;
-  margin-right: 0;
-  margin-left: 4px;
-}
-#cssmenu.align-right ul li:first-child,
-#cssmenu.align-right ul li:first-child > a {
-  border-bottom-right-radius: 3px;
-}
-#cssmenu #bg-one,
-#cssmenu #bg-two,
-#cssmenu #bg-three,
-#cssmenu #bg-four {
-  position: absolute;
-  bottom: 0;
-  width: 100%;
-  border-bottom-left-radius: 3px;
-  border-bottom-right-radius: 3px;
-}
-#cssmenu #bg-one {
-  height: 10px;
-  background: #0f71ba;
-}
-#cssmenu #bg-two {
-  height: 50px;
-  z-index: 2;
-}
-#cssmenu #bg-three {
-  bottom: 4px;
-  height: 50px;
-  z-index: 3;
-  background: #222222;
-  background: -moz-linear-gradient(top, #555555 0%, #222222 100%);
-  background: -webkit-gradient(linear, left top, left bottom, color-stop(0%, #555555), color-stop(100%, #222222));
-  background: -webkit-linear-gradient(top, #555555 0%, #222222 100%);
-  background: -o-linear-gradient(top, #555555 0%, #222222 100%);
-  background: -ms-linear-gradient(top, #555555 0%, #222222 100%);
-  background: linear-gradient(to bottom, #555555 0%, #222222 100%);
-}
-#cssmenu #bg-four {
-  bottom: 4px;
-  height: 50px;
-  z-index: 4;
-  background: url('images/bg.png');
-}
-#cssmenu ul {
-  height: 50px;
-}
-#cssmenu ul li {
-  float: left;
-  margin-right: 4px;
-  border-top-left-radius: 3px;
-  border-top-right-radius: 3px;
-  z-index: 5;
-}
-#cssmenu ul li a {
-  padding: 24px 30px 20px 30px;
-  margin-bottom: 4px;
-  border-top-left-radius: 3px;
-  border-top-right-radius: 3px;
-  color: #eeeeee;
-  font-size: 15px;
-  text-decoration: none;
-}
-#cssmenu ul li:first-child,
-#cssmenu ul li:first-child > a {
-  border-bottom-left-radius: 3px;
-}
-#cssmenu ul li:hover,
-#cssmenu ul li.active {
-  background: #0f71ba;
-  background: -moz-linear-gradient(top, #3fa4f0 0%, #0f71ba 100%);
-  background: -webkit-gradient(linear, left top, left bottom, color-stop(0%, #3fa4f0), color-stop(100%, #0f71ba));
-  background: -webkit-linear-gradient(top, #3fa4f0 0%, #0f71ba 100%);
-  background: -o-linear-gradient(top, #3fa4f0 0%, #0f71ba 100%);
-  background: -ms-linear-gradient(top, #3fa4f0 0%, #0f71ba 100%);
-  background: linear-gradient(to bottom, #3fa4f0 0%, #0f71ba 100%);
-}
-#cssmenu ul li a:hover,
-#cssmenu ul li.active > a {
-  color: #ffffff;
-}
-
-.about{
-  text-align: center;
-}
-#aheading{
-  text-decoration: underline;
-  text-align: left;
-
-}
-
-h3, h4, h2 {text-align: center}
