--- UI/entry_content.html
@@ -1,65 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>MyDiary | Entry View</title>
-    <link href='https://fonts.googleapis.com/css?family=Titillium+Web:400,300,600' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" type="text/css" href="css/navbar.css">
-    <link rel="stylesheet" type="text/css" href="css/entries.css">
-    <link rel="stylesheet" type="text/css" href="css/buttons.css">
-    <link rel="stylesheet" type="text/css" href="css/normalize.min.css">
-    <script src="js/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="js/navbar.js" type="text/javascript"></script>
-
-
-    <!--Navigation Bar-->
-    <div id='cssmenu'>
-        <ul>
-            <li><a href='index.html'><span>MyDiary</span></a></li>
-            <li class='active'><a href='entries.html'><span>Entries</span></a></li>
-            <li><a href='profile.html'><span>Profile</span></a></li>
-            <li class='last'><a href='index.html'><span> | Sign Out</span></a></li>
-        </ul>
-    </div>
-
-</head>
-
-
-<body background="images/bg.png">
-
-<div class="DivTable" id="EntryView">
-    <div class="Title">
-        <p>View Entry</p>
-    </div>
-
-    <div class="row" id="subtitle">
-        <p>1. &emsp;Andela PreBootCamp</p>
-    </div>
-    <div class="row" id="EntryCategory">
-        <p>Career</p>
-    </div>
-
-    <div class="row">
-        <p>
-            So there I was, Lorem ipsum dolor sit amet consectetur adipisicing elit.
-            Quidem accusamus provident ad id culpa nobis dicta accusantium ipsa magnam facere reiciendis,
-            quos voluptate eligendi numquam.Lorem ipsum dolor sit amet consectetur adipisicing elit.
-            Quidem accusamus provident ad id culpa nobis dicta accusantium
-        </p>
-    </div>
-    <div class="row">
-        <p>You Created this on 14/06/2018</p>
-    </div>
-    <div class="row">
-        <p>Last Updated on 16/06/2018</p>
-    </div>
-    <div class="row">
-        <button class="button" onclick="location.href='entries.html';"> << Back</button>
-        <span class="btnEnd"><button class="button" onclick="location.href='modify_entry.html';">Modify</button></span>
-    </div>
-
-</div>
-
-
-</body>
-</html>
