--- UI/index.html
@@ -1,121 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>MyDiary</title>
-    <link href='https://fonts.googleapis.com/css?family=Titillium+Web:400,300,600' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" href="css/normalize.min.css">
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="stylesheet" href="css/buttons.css">
-    <script src="js/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="js/navbar.js" type="text/javascript"></script>
-</head>
-
-<body background="images/bg.png">
-
-<!--Navigation Bar-->
-<div id='cssmenu'>
-    <ul>
-        <li class='active'><a href='#'><span>MyDiary</span></a></li>
-        <li><a href='entries.html'><span>Entries</span></a></li>
-        <li><a href='profile.html'><span>Profile</span></a></li>
-        <li class='last'><a href='about.html'><span>About</span></a></li>
-    </ul>
-</div>
-
-<h3>Pen Down Your Thoughts & Feelings Freely For Free with MyDiary.</h3>
-<h4>Online Journal</h4>
-
-<div class="form">
-
-    <ul class="tab-group">
-        <li class="tab active"><a href="#signup">Sign Up</a></li>
-        <li class="tab"><a href="#login">Log In</a></li>
-    </ul>
-
-    <div class="tab-content">
-        <div id="signup">
-            <h1>Sign Up for Free</h1>
-
-            <form action="entries.html" method="post">
-
-                <div class="top-row">
-                    <div class="field-wrap">
-                        <label>
-                            First Name<span class="req" for="fName">*</span>
-                        </label>
-                        <input type="text" id="fName" required autocomplete="off"/>
-                    </div>
-
-                    <div class="field-wrap">
-                        <label>
-                            Last Name<span class="req" for="lName">*</span>
-                        </label>
-                        <input type="text" id="lName" required autocomplete="off"/>
-                    </div>
-                </div>
-
-                <div class="field-wrap">
-                    <label>
-                        Email Address<span class="req" for="userEmail">*</span>
-                    </label>
-                    <input type="email" id="newUserEmail" required autocomplete="off"/>
-                </div>
-
-                <div class="field-wrap">
-                    <label>
-                        Set A Password<span class="req" for="passphrase1">*</span>
-                    </label>
-                    <input type="password" id="passphrase1" required autocomplete="off"/>
-                </div>
-
-                <div class="field-wrap">
-                    <label>
-                        Confirm Password<span class="req" for="passphrase2">*</span>
-                    </label>
-                    <input type="password" id="passphrase2" required autocomplete="off"/>
-                </div>
-
-                <span class="btnCenter"><button type="submit" class="button">Get Started</button></span>
-
-            </form>
-
-        </div>
-
-        <div id="login">
-            <h1>Log In</h1>
-
-            <form action="entries.html" method="post">
-
-                <div class="field-wrap">
-                    <label>
-                        Email Address<span class="req">*</span>
-                    </label>
-                    <input type="email" id="userEmail" required autocomplete="off"/>
-                </div>
-
-                <div class="field-wrap">
-                    <label>
-                        Password<span class="req">*</span>
-                    </label>
-                    <input type="password" id="passphrase" required autocomplete="off"/>
-                </div>
-
-                <p class="forgot"><a href="#">Forgot Password?</a></p>
-
-                <button class="button">Log In</button>
-
-            </form>
-
-        </div>
-
-    </div><!-- tab-content -->
-
-</div> <!-- /form -->
-<script src='js/jquery.min.js'></script>
-<script src="js/index.js"></script>
-
-</body>
-</html>
