--- UI/js/navbar.js
@@ -1,5 +1,0 @@
-( function( $ ) {
-$( document ).ready(function() {
-$('#cssmenu').prepend('<div id="bg-one"></div><div id="bg-two"></div><div id="bg-three"></div><div id="bg-four"></div>');
-});
-} )( jQuery );
