--- UI/modify_entry.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns="http://www.w3.org/1999/html">
-
-<head>
-    <link rel="stylesheet" type="text/css" href="UI/css/entries.css"/>
-    <link rel="stylesheet" type="text/css" href="UI/css/entries.css"/>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>MyDiary</title>
-    <link href='https://fonts.googleapis.com/css?family=Titillium+Web:400,300,600' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" href="css/normalize.min.css">
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="stylesheet" href="css/buttons.css">
-    <script src="js/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="js/navbar.js" type="text/javascript"></script>
-
-    <!--Navigation Bar-->
-    <div id='cssmenu'>
-        <ul>
-            <li><a href='index.html'><span>MyDiary</span></a></li>
-            <li class='active'><a href='entries.html'><span>Entries</span></a></li>
-            <li><a href='profile.html'><span>Profile</span></a></li>
-            <li class='last'><a href='index.html'><span> | Sign Out</span></a></li>
-        </ul>
-    </div>
-</head>
-
-<body background="images/bg.png">
-
-<div class="form">
-
-    <ul class="tab-group">
-
-    </ul>
-
-    <div class="tab-content">
-        <div id="editEntty">
-            <h1>Modify Entry</h1>
-                <form>
-
-                    <div class="field-wrap">
-                        <label>
-                            Category <span class="req">*</span>
-                        </label>
-
-                        <input list="category" name="category" >
-                        <datalist id="category">
-                            <option value="Career">
-                            <option value="Life">
-                            <option value="Love">
-                            <option value="Entertainment">
-                            <option value="Relationships">
-                            <option value="Family">
-                            <option value="Travel">
-                        </datalist>
-                    </div>
-
-                    <div class="field-wrap">
-                        <label>
-                            Title <span class="req">*</span>
-                        </label>
-                        <input type="text" id="entryTitle" name="title" required/>
-                    </div>
-
-                    <div class="field-wrap">
-                        <label>
-                            Content <span class="req">*</span>
-                        </label>
-                        <textarea name="entry_content" cols="50" rows="8">So there I was, Lorem ipsum dolor sit amet consectetur adipisicing elit.
-            Quidem accusamus provident ad id culpa nobis dicta accusantium ipsa magnam facere reiciendis,
-            quos voluptate eligendi numquam.Lorem ipsum dolor sit amet consectetur adipisicing elit.
-            Quidem accusamus provident ad id culpa nobis dicta accusantium</textarea>
-                    </div>
-
-
-                    <button type="submit" class="button" onclick="location.href='entries.html'">Update Entry</button>
-                    <span class="btnSpaced">
-                        <button class="button" onclick="location.href='entries.html';">Discard Changes</button>
-                        <button class="button" id="btnDelete" onclick="confirm('Are you sure you want to delete this Entry?')">Delete Entry</button>
-                    </span>
-                </form>
-
-        </div>
-
-        <div></div>
-
-    </div><!-- tab-content -->
-
-</div> <!-- /form -->
-<script src='js/jquery.min.js'></script>
-<script src="js/index.js"></script>
-
-</body>
-</html>
