--- UI/profile.html
@@ -1,70 +1,0 @@
-<html>
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>MyDiary | Profile</title>
-    <link href='https://fonts.googleapis.com/css?family=Titillium+Web:400,300,600' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" type="text/css" href="css/navbar.css">
-    <link rel="stylesheet" type="text/css" href="css/profile.css">
-    <link rel="stylesheet" type="text/css" href="css/buttons.css">
-    <link rel="stylesheet" type="text/css" href="css/normalize.min.css">
-    <script src="js/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="js/navbar.js" type="text/javascript"></script>
-
-    <!--Navigation Bar-->
-    <div id='cssmenu'>
-        <ul>
-            <li><a href='index.html'><span>MyDiary</span></a></li>
-            <li><a href='entries.html'><span>Entries</span></a></li>
-            <li class='active'><a href='profile.html'><span>Profile</span></a></li>
-            <li class='last'><a href='index.html'><span> | Sign Out</span></a></li>
-        </ul>
-    </div>
-</head>
-
-<body background="images/bg.png">
-
-
-<div id="DivContainer">
-    <div class="ProfileDiv" id="userProf">
-        <p>Byron Taaka</p>
-        <img src="images/profile.png" alt="Profile Picture"/>
-
-
-        <button class="ProfileDiv" id="btnProfPic"
-                onclick="alert('Not Database exists so far to store your profile picture')">Change Avatar
-        </button>
-
-    </div>
-    <div class="ProfileDiv" id="userDetails">
-        <br/>
-
-        Nick: &emsp;&ensp; Byron<br/>
-        Email: &emsp; 4dbyron@gmail.com<br/>
-        Joined: &emsp;9/07/2018<br/>
-
-        <p><br/>
-            <b>8</b> entries created since Joining<br/>
-            <b>7</b> entries currently active<br/>
-        </p>
-
-        <button class="ProfileDiv" id="btnPassword"
-                onclick="location.href='update_password.html'">Change Password
-        </button>
-
-    </div>
-    <div class="ProfileDiv" id="userSettings"><p></p><br/>
-        Daily Reminder:
-        <span class="notifySpace">
-            <label class="notify"><input type="checkbox"><span class="reminder round"></span></label>
-        </span>
-    </div>
-</div>
-</div>
-
-
-<script src='js/jquery.min.js'></script>
-<script src="js/index.js"></script>
-
-</body>
-</html>
