--- UI/update_password.html
@@ -1,85 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns="http://www.w3.org/1999/html">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>MyDiary</title>
-    <link href='https://fonts.googleapis.com/css?family=Titillium+Web:400,300,600' rel='stylesheet' type='text/css'>
-    <link rel="stylesheet" href="css/normalize.min.css">
-    <link rel="stylesheet" href="css/style.css">
-    <link rel="stylesheet" href="css/buttons.css">
-    <script src="js/jquery-latest.min.js" type="text/javascript"></script>
-    <script src="js/navbar.js" type="text/javascript"></script>
-
-</head>
-
-<body background="images/bg.png">
-
-<!--Navigation Bar-->
-<div id='cssmenu'>
-    <ul>
-        <li><a href='#'><span>MyDiary</span></a></li>
-        <li><a href='entries.html'><span>Entries</span></a></li>
-        <li class='active'><a href='profile.html'><span>Profile</span></a></li>
-        <li class='last'><a href='about.html'><span>About</span></a></li>
-    </ul>
-</div>
-
-
-<!-- The Modal -->
-<div id="myModal" class="modal">
-
-    <!-- Modal content -->
-    <div class="modal-content">
-        <span class="close">&times;</span>
-        <div class="form">
-            <h1>Change Password</h1>
-
-            <form action="profile.html" method="post">
-
-                <div>
-                    <div class="field-wrap">
-                        <label>
-                            Enter Old Password<span class="req" for="passphrase1">*</span>
-                        </label>
-                        <input type="password" id="oldPassPhrase" required autocomplete="off"/>
-                    </div>
-
-                    <div class="field-wrap">
-                        <label>
-                            Set New Password<span class="req" for="passphrase1">*</span>
-                        </label>
-                        <input type="password" id="newPassPhrase" required autocomplete="off"/>
-                    </div>
-
-                    <div class="field-wrap">
-                        <label>
-                            Confirm Password<span class="req" for="newPassPhrase2">*</span>
-                        </label>
-                        <input type="password" id="newPassPhrase2" required autocomplete="off"/>
-                    </div>
-
-                </div>
-                <button class="button" onclick="location.href='profile.html'">Cancel</button>
-                &emsp;&emsp;&emsp;&emsp;
-                <span class="btnCenter"><button type="submit" class="button">Update Password</button></span>
-            </form>
-            <p/>
-
-        </div> <!-- /form -->
-    </div>
-</div>
-
-
-<script type="text/javascript" src="js/form_password.js"></script>
-<script type="text/javascript" src='js/jquery.min.js'></script>
-<script type="text/javascript" src="js/index.js"></script>
-
-</body>
-</html>
-
-<body>
-
-</body>
-</html>
